% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MelSign_df_Karras.R
\docType{data}
\name{MelSign_df_Karras}
\alias{MelSign_df_Karras}
\title{MelSign gene signature}
\format{
A data frame with 198 rows and 2 variables:
\describe{
\item{gene}{Gene symbol}
}
}
\source{
Karras et al.
}
\usage{
MelSign_df_Karras
}
\description{
Dataset containing the MelSign signature from Karras et al.
}
\keyword{datasets}
